Firts commit from Local Computer 42

+++ Libft/ft_strdup.c
@@ -6,14 +6,16 @@
 /*   By: nbiescas <nbiescas@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/02/02 16:52:00 by nbiescas          #+#    #+#             */
-/*   Updated: 2023/02/06 16:03:49 by nbiescas         ###   ########.fr       */
+/*   Updated: 2023/03/07 10:43:42 by nbiescas         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
+
+#include "libft.h"
 
 char	*ft_strdup(const char *s1)
 {
 	char	*str;
-	int		len;
+	size_t	len;
 
 	len = ft_strlen(s1);
 	str = malloc((len + 1) * sizeof(char));
